+++ CHANGELOG.md
@@ -3,14 +3,43 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v2.2.2dev - 2023-08-29
+## v2.4.0dev [date]
 
 ### `Changed`
 
-- [203](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.9`
-- [203](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.8`
+- [#247](https://github.com/nf-core/epitopeprediction/pull/247) - Update to nf-core template `3.0.2`
+
+## v2.3.1 - Oesterberg - 2024-05-17
+
+### `Changed`
+
+- [#243](https://github.com/nf-core/epitopeprediction/pull/243) - Update to nf-core template `2.14.1`
+- [#237](https://github.com/nf-core/epitopeprediction/pull/237) - Update to nf-core template `2.13.1`
+
+### `Fixed`
+
+- [#243](https://github.com/nf-core/epitopeprediction/pull/243) - Add check for protein map to prevent failure if no information is available
+
+## v2.3.0 - Oesterberg - 2024-02-26
+
+### `Changed`
+
+- [#233](https://github.com/nf-core/epitopeprediction/pull/233) - Update to nf-core template `2.13`
+- [#228](https://github.com/nf-core/epitopeprediction/pull/228) - Update to nf-core template `2.12`
+- [#227](https://github.com/nf-core/epitopeprediction/pull/227) Prevent crash if no transcript is found (in splitted vcf)
+- [#220](https://github.com/nf-core/epitopeprediction/pull/220) - Switch to nf-validation to parse samplesheet
+- [#213](https://github.com/nf-core/epitopeprediction/pull/213) - Update epytope and Ensembl reference handling and update to nf-core template `2.10`
 - [#206](https://github.com/nf-core/epitopeprediction/issues/206) - Update the row checker class.
-- [#205](https://github.com/nf-core/epitopeprediction/issues/205) - Update general structure of MHC binding subworkflow
+- [#203](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.9`, rename param `genome_version` to `genome_reference`, add functionality to handle BioMart archive urls
+
+### `Fixed`
+
+- [#219](https://github.com/nf-core/epitopeprediction/pull/219) - Fix `EXTERNAL_TOOLS_IMPORT`` container registry and bump version
+- [#227](https://github.com/nf-core/epitopeprediction/pull/227) - Prevent crash if no transcript is found (in splitted vcf)
+
+### `Removed`
+
+- [#221](https://github.com/nf-core/epitopeprediction/pull/221) - Remove support of `GSvar` and variant `tsv` input files
 
 ## v2.2.1 - WaldhaeuserOst Hotfix - 2023-03-16